photo_trans/requirements.txt
ffmpeg-python==4.5.1
pyheif==0.6.0
Pillow==8.3.1
pyexiv2==0.3.2
